chore: Remove regex-based CParser and make pycparser required
Consolidate C parsing to use only pycparser-based implementation.

Changes:
- Delete regex-based c_parser.py and CParserPyCParser class
- Rename c_parser_pycparser.py → c_parser.py
- Rename class CParserPyCParser → CParser
- Remove optional import logic (pycparser now required)
- Update pyproject.toml: pycparser in core dependencies
- Remove [parser] optional dependency group
- Update FunctionDatabase to use pycparser exclusively
- Clean up test files (rename/remove old test files)
- Delete demo_parser_comparison.py script

Quality:
- isort: ✅ Pass
- Ruff: ✅ Pass
- Mypy: ✅ Pass
- Pytest: ✅ 354 passed, 11 skipped
- Coverage: 90%

Closes #56
